Move filter_chain stuff into run_policy namespace and its respective directory
It is a thing for run-policies only.
