+++ bandit/run_policies/filter_chain.h
@@ -1,11 +1,11 @@
-#ifndef BANDIT_FILTER_CHAIN_H
-#define BANDIT_FILTER_CHAIN_H
+#ifndef BANDIT_RUN_POLICIES_FILTER_CHAIN_H
+#define BANDIT_RUN_POLICIES_FILTER_CHAIN_H
 
 #include <string>
 #include <vector>
 
 namespace bandit {
-  namespace detail {
+  namespace run_policy {
     struct filter_chain_element {
       std::string pattern;
       bool skip;
